Removing deprecated notebooks
Changes:
- Removed deprecated notebooks
